Move statistics back to monitors/stats
