+++ MIGRATION.md
@@ -1,4 +1,5 @@
 # 0.15.0 -> 0.16.0
+
 - `EventManager` is refactored to avoid calling function from `Fuzzer`, thus we do not evaluate testcases in `EventManager` anymore.
   - Now we have `EventReceiver` in `events` module, and `EventProessor` in `fuzzer` module.
   - `EventReceiver` is responsible for receiving testcases and delegates its evaluation to `EventProcessor`.
@@ -6,13 +7,13 @@
   - Since we don't evaluate testcases in the `EventManager` anymore. `on_fire` and `post_exec` have been deleted from `EventManagerHook`.
   - Similarly `pre_exec` has been renamed to `pre_receive`.
 - `AsanModule` now uses a `builder()` method for constructing its instances.
-- `Monitor` is refactored. Most statistics have been extracted into an individual module `statistics` at crate root, and `Monitor` is kept as display only.
+- `Monitor` is refactored. Most statistics have been extracted into an individual `stats` module under `monitors`.
   - There is a `ClientStatsManager` to manage client statistics, and is owned by `EventManager`. Most of previous `Monitor`'s trait methods have been moved to the `ClientStatsManager`.
-  - `user_monitor` has been renamed to `user_stats`, and its structure definitions have been moved to `statistics` module as well.
-  - `introspection_monitor` has been renamed to `introspection_stats`, and perf-related structure definitions have been renamed and moved to `statistics` module as well.
+  - `user_monitor` has been renamed to `user_stats`, `introspection_monitor` has been renamed to `introspection_stats`, perf-related structure definitions have been renamed, and all were moved to the `stats` module.
   - `OnDiskTomlMonitor`, `OnDiskJsonMonitor`, `OnDiskJsonAggregateMonitor` are now no longer takes a base monitor to wrap. If you want to use multiple monitors together, use `libafl::combine_monitor!`.
 
 # 0.14.1 -> 0.15.0
+
 - `MmapShMem::new` and `MmapShMemProvider::new_shmem_with_id` now take `AsRef<Path>` instead of a byte array for the filename/id.
 - The closure passed to a `DumpToDiskStage` now provides the `Testcase` instead of just the `Input`.
 - `StatsStage` is deleted, and it is superceded by `AflStatsStage`
@@ -35,8 +36,10 @@
   - Related: removed `hash_simple` from `MapObserver`
 
 # 0.14.0 -> 0.15.0
+
 - Removed `with_observers` from `Executor` trait.
 - `MmapShMemProvider::new_shmem_persistent` has been removed in favour of `MmapShMem::persist`. You probably want to do something like this: `let shmem = MmapShMemProvider::new()?.new_shmem(size)?.persist()?;`
 
 # Pre 0.9 -> 0.9
+
 - [Migrating from LibAFL <0.9 to 0.9](https://aflplus.plus/libafl-book/design/migration-0.9.html)
