+++ libafl/src/events/mod.rs
@@ -13,6 +13,8 @@ pub use centralized::*;
 #[cfg(feature = "std")]
 pub mod launcher;
 
+pub mod stats;
+
 pub mod llmp;
 pub use llmp::*;
 #[cfg(feature = "tcp_manager")]
@@ -45,10 +47,10 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::{
+    events::stats::user_stats::UserStats,
     executors::ExitKind,
     inputs::Input,
     state::{HasExecutions, HasLastReportTime, MaybeHasClientPerfMonitor},
-    statistics::user_stats::UserStats,
     Error, HasMetadata,
 };
 
@@ -107,7 +109,7 @@ pub struct EventManagerId(
 #[cfg(all(unix, feature = "std", feature = "multi_machine"))]
 use crate::events::multi_machine::NodeId;
 #[cfg(feature = "introspection")]
-use crate::statistics::perf_stats::ClientPerfStats;
+use crate::events::stats::perf_stats::ClientPerfStats;
 use crate::{observers::TimeObserver, stages::HasCurrentStageId};
 
 /// The log event severity
