+++ libafl/src/events/simple.rs
@@ -25,20 +25,19 @@ use serde::Serialize;
 use super::{std_on_restart, AwaitRestartSafe, ProgressReporter, RecordSerializationTime};
 #[cfg(all(unix, feature = "std", not(miri)))]
 use crate::events::EVENTMGR_SIGHANDLER_STATE;
+#[cfg(feature = "std")]
+use crate::{events::stats::ClientStats, monitors::SimplePrintingMonitor, state::HasSolutions};
 use crate::{
     events::{
-        std_maybe_report_progress, std_report_progress, BrokerEventResult, CanSerializeObserver,
-        Event, EventFirer, EventManagerId, EventReceiver, EventRestarter, HasEventManagerId,
-        SendExiting,
+        stats::manager::ClientStatsManager, std_maybe_report_progress, std_report_progress,
+        BrokerEventResult, CanSerializeObserver, Event, EventFirer, EventManagerId, EventReceiver,
+        EventRestarter, HasEventManagerId, SendExiting,
     },
     monitors::Monitor,
     stages::HasCurrentStageId,
     state::{HasExecutions, HasLastReportTime, MaybeHasClientPerfMonitor, Stoppable},
-    statistics::manager::ClientStatsManager,
     Error, HasMetadata,
 };
-#[cfg(feature = "std")]
-use crate::{monitors::SimplePrintingMonitor, state::HasSolutions, statistics::ClientStats};
 
 /// The llmp connection from the actual fuzzer to the process supervising it
 const _ENV_FUZZER_SENDER: &str = "_AFL_ENV_FUZZER_SENDER";
