+++ libafl/src/lib.rs
@@ -79,7 +79,6 @@ pub mod observers;
 pub mod schedulers;
 pub mod stages;
 pub mod state;
-pub mod statistics;
 
 pub use fuzzer::*;
 pub use libafl_bolts::{nonzero, Error};
@@ -91,8 +90,7 @@ pub mod prelude {
 
     pub use super::{
         corpus::*, events::*, executors::*, feedbacks::*, fuzzer::*, generators::*, inputs::*,
-        monitors::*, mutators::*, observers::*, schedulers::*, stages::*, state::*, statistics::*,
-        *,
+        monitors::*, mutators::*, observers::*, schedulers::*, stages::*, state::*, *,
     };
 }
