+++ libafl/src/monitors/stats/manager.rs
