+++ libafl/src/stages/tracing.rs
@@ -9,7 +9,7 @@ use core::{fmt::Debug, marker::PhantomData};
 use libafl_bolts::Named;
 
 #[cfg(feature = "introspection")]
-use crate::statistics::perf_stats::PerfFeature;
+use crate::events::stats::perf_stats::PerfFeature;
 use crate::{
     corpus::HasCurrentCorpusId,
     executors::{Executor, HasObservers, ShadowExecutor},
