+++ libafl/src/state/mod.rs
@@ -27,7 +27,7 @@ mod stack;
 pub use stack::StageStack;
 
 #[cfg(feature = "introspection")]
-use crate::statistics::perf_stats::ClientPerfStats;
+use crate::events::stats::perf_stats::ClientPerfStats;
 use crate::{
     corpus::{Corpus, CorpusId, HasCurrentCorpusId, HasTestcase, InMemoryCorpus, Testcase},
     events::{Event, EventFirer, LogSeverity},
